CSHARP-2226: Refactor so there are no public API changes.
