+++ tests/MongoDB.Driver.Legacy.Tests/MongoClientTests.cs
@@ -13,16 +13,12 @@
 * limitations under the License.
 */
 
-using System;
-using System.Linq;
-using System.Reflection;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using FluentAssertions;
-using MongoDB.Bson.TestHelpers;
+using MongoDB.Bson;
 using MongoDB.Driver;
-using MongoDB.Driver.Core.Clusters;
-using MongoDB.Driver.Core.Clusters.ServerSelectors;
-using MongoDB.Driver.Core.Servers;
-using MongoDB.Driver.Core.TestHelpers;
+using MongoDB.Driver.Core.Operations;
 using Xunit;
 
 namespace MongoDB.Driver.Tests
@@ -60,65 +56,5 @@ namespace MongoDB.Driver.Tests
 
             Assert.Same(server1, server2);
         }
-    }
-
-    public class AreSessionsSupportedServerSelectorTests
-    {
-        [Theory]
-        [InlineData("{ clusterType : 'Standalone', servers : [ { state : 'Disconnected', type : 'Unknown' } ]}")]
-        public void SelectServers_should_set_ClusterDescription(string clusterDescriptionJson)
-        {
-            var subject = CreateSubject();
-            var cluster = ClusterDescriptionParser.Parse(clusterDescriptionJson);
-            var connectedServers = cluster.Servers.Where(s => s.State == ServerState.Connected);
-
-            var result = subject.SelectServers(cluster, connectedServers);
-
-            AreSessionsSupportedServerSelectorReflector.ClusterDescription(subject).Should().BeSameAs(cluster);
-        }
-
-        [Theory]
-        [InlineData("{ connectionMode : 'Direct', clusterType : 'ReplicaSet', servers : [ { state : 'Connected', type : 'ReplicaSetArbiter' } ]}")]
-        public void SelectServers_should_return_all_servers_when_connection_mode_is_direct(string clusterDescriptionJson)
-        {
-            var subject = CreateSubject();
-            var cluster = ClusterDescriptionParser.Parse(clusterDescriptionJson);
-            var connectedServers = cluster.Servers.Where(s => s.State == ServerState.Connected).ToList();
-
-            var result = subject.SelectServers(cluster, connectedServers);
-
-            result.Should().Equal(connectedServers);
-        }
-
-        [Theory]
-        [InlineData("{ clusterType : 'ReplicaSet', servers : [ { state : 'Connected', type : 'ReplicaSetArbiter' }, { state : 'Connected', type : 'ReplicaSetPrimary' } ]}")]
-        public void SelectServers_should_return_data_bearing_servers_when_connection_mode_is__not_direct(string clusterDescriptionJson)
-        {
-            var subject = CreateSubject();
-            var cluster = ClusterDescriptionParser.Parse(clusterDescriptionJson);
-            var connectedServers = cluster.Servers.Where(s => s.State == ServerState.Connected).ToList();
-            var dataBearingServers = connectedServers.Skip(1).Take(1);
-
-            var result = subject.SelectServers(cluster, connectedServers);
-
-            result.Should().Equal(dataBearingServers);
-        }
-
-        // private methods
-        private IServerSelector CreateSubject()
-        {
-            return AreSessionsSupportedServerSelectorReflector.CreateInstance();
-        }
-    }
-
-    public static class AreSessionsSupportedServerSelectorReflector
-    {
-        public static IServerSelector CreateInstance()
-        {
-            var type = typeof(MongoClient).GetTypeInfo().Assembly.GetType("MongoDB.Driver.MongoClient+AreSessionsSupportedServerSelector");
-            return (IServerSelector)Activator.CreateInstance(type);
-        }
-
-        public static ClusterDescription ClusterDescription(IServerSelector obj) => (ClusterDescription)Reflector.GetFieldValue(obj, nameof(ClusterDescription), BindingFlags.Public | BindingFlags.Instance);
     }
 }
